requirements.txt
bokeh==1.2.0
certifi==2019.9.11
chardet==3.0.4
Django==2.0.6
django-allow-cidr==0.3.0
idna==2.8
influxdb==5.2.3
Jinja2==2.10.3
MarkupSafe==1.1.1
netaddr==0.7.19
numpy==1.17.2
packaging==19.2
paho-mqtt==1.4.0
pandas==0.25.1
Pillow==7.1.2
psycopg2==2.8.5
pudb==2019.1
Pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2018.4
PyYAML==5.1.2
requests==2.22.0
rfc3339==6.2
six==1.12.0
sqlparse==0.2.4
tornado==6.0.3
urllib3==1.25.6
urwid==2.0.1
